+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "samseg/samseg/hdav"]
-	path = samseg/samseg/hdav
+[submodule "python/freesurfer/samseg/hdav"]
+	path = python/freesurfer/samseg/hdav
 	url = https://github.com/innolitics/hdav.git
