+++ python/freesurfer/samseg/Testing/test_show_figures.py
