+++ python/freesurfer/samseg/__init__.py
@@ -1,0 +1,4 @@
+from .utilities import Specification, kvlReadCompressionLookupTable, kvlReadSharedGMMParameters
+from .register_atlas import registerAtlas
+from .samsegment import samsegment
+from .figures import initVisualizer
