+++ python/freesurfer/samseg/samsegment.py
@@ -1,23 +1,25 @@
-import logging
-import os.path
+import os
 
-from samseg.dev_utils.debug_client import run_test_cases, create_checkpoint_manager, load_starting_fixture
-from samseg.run_utilities import Specification
-from samseg.samseg_ported_part1 import samsegment_part1
-from samseg.samseg_ported_part2 import samsegment_part2
-from samseg.samseg_ported_part3 import samsegment_part3
+from .utilities import Specification
+from .samsegment_part1 import samsegment_part1
+from .samsegment_part2 import samsegment_part2
+from .samsegment_part3 import samsegment_part3
 
-logger = logging.getLogger(__name__)
+# from .dev_utils.debug_client import run_test_cases, create_checkpoint_manager, load_starting_fixture
+
 
 def samsegment(
-        imageFileNames,
-        transformedTemplateFileName,
-        modelSpecifications,
-        optimizationOptions,
-        savePath,
-        visualizer,
-        checkpoint_manager=None
+    imageFileNames,
+    transformedTemplateFileName,
+    modelSpecifications,
+    optimizationOptions,
+    savePath,
+    visualizer,
+    checkpoint_manager=None
 ):
+
+    # ------ Setup ------
+
     part0_results_dict = Specification({
         'imageFileNames': imageFileNames,
         'transformedTemplateFileName': transformedTemplateFileName,
@@ -26,9 +28,13 @@ def samsegment(
         'savePath': savePath,
         'visualizer': visualizer,
     })
+
     if checkpoint_manager:
         checkpoint_manager.save_specification(part0_results_dict, 'part0', 1)
-    logger.info('calling part1...')
+
+    # ------ Samsegment Part 1 ------
+
+    print('calling part1...')
     part1_results_dict = samsegment_part1(
         imageFileNames,
         transformedTemplateFileName,
@@ -38,10 +44,13 @@ def samsegment(
         visualizer,
         checkpoint_manager
     )
+
     if checkpoint_manager:
         checkpoint_manager.save(part1_results_dict, 'part1', 1)
-    logger.info('calling part2...')
+    
+    # ------ Samsegment Part 2 ------
 
+    print('calling part2...')
     part2_results_dict = samsegment_part2(
         modelSpecifications,
         optimizationOptions,
@@ -49,10 +58,13 @@ def samsegment(
         visualizer,
         checkpoint_manager
     )
+
     if checkpoint_manager:
         checkpoint_manager.save(part2_results_dict, 'part2', 1)
+    
+    # ------ Samsegment Part 3 ------
 
-    logger.info('calling part3...')
+    print('calling part3...')
     part3_results_dict = samsegment_part3(
         modelSpecifications,
         optimizationOptions,
@@ -62,9 +74,13 @@ def samsegment(
         visualizer,
         checkpoint_manager
     )
+
     if checkpoint_manager:
         checkpoint_manager.save(part3_results_dict, 'part3', 1)
-    logger.info('...all parts completed')
+
+    # ------ Done ------
+
+    print('...all parts completed')
     names = part1_results_dict['names']
     FreeSurferLabels = part3_results_dict['FreeSurferLabels']
     volumesInCubicMm = part3_results_dict['volumesInCubicMm']
@@ -76,7 +92,7 @@ def samsegment(
     return [FreeSurferLabels, names, volumesInCubicMm]
 
 
-def test_samseg_ported(case_name, case_file_folder, savePath):
+def test_samsegment(case_name, case_file_folder, savePath):
     checkpoint_manager = create_checkpoint_manager(case_file_folder)
     fixture = load_starting_fixture()
     results = samsegment(
@@ -92,4 +108,4 @@ def test_samseg_ported(case_name, case_file_folder, savePath):
 
 
 if __name__ == '__main__':
-    run_test_cases(action=test_samseg_ported)
+    run_test_cases(action=test_samsegment)
