+++ samseg/CMakeLists.txt
@@ -1,11 +1,3 @@
-project(pysurfer)
-
-install(CODE "
-  message(STATUS \"Installing: ${CMAKE_INSTALL_PREFIX}/python/samseg\")
-  execute_process(COMMAND bash -c \"pip install ${CMAKE_CURRENT_SOURCE_DIR} --upgrade --target=${CMAKE_INSTALL_PREFIX}/python\" RESULT_VARIABLE retcode)
-  if(NOT \${retcode} STREQUAL 0)
-    message(FATAL_ERROR \"Could not install samseg python module\")
-  endif()"
-)
+project(samseg)
 
 install(PROGRAMS run_samseg sbtiv DESTINATION bin)
