--- samseg/samseg/samseg_utilities.py
@@ -1,3 +1,0 @@
-def intracranial_volume(structures, includeStructures):
-    # sum over included structure volumes
-    return sum(structure['vol'] for structure in structures if structure['name'] in includeStructures)
